Update changes and create new test classes

+++ src/main/java/com/banking/account/transact/domain/transaction/ComputeTransactionRepository.java
@@ -1,5 +1,7 @@
 package com.banking.account.transact.domain.transaction;
 
+import java.math.BigDecimal;
+
 public interface ComputeTransactionRepository {
-    public double computeTaxTransaction(double valor);
+    public BigDecimal computeTaxTransaction(BigDecimal valor);
 }
